% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mongo_conn}
\alias{get_mongo_conn}
\title{Get or Create a MongoDB Connection}
\usage{
get_mongo_conn(db, collection, uri)
}
\arguments{
\item{db}{A string specifying the MongoDB database name.}

\item{collection}{A string specifying the MongoDB collection name.}

\item{uri}{A string specifying the MongoDB connection URI.}
}
\value{
A `mongolite::mongo` connection object.
}
\description{
This function retrieves a cached MongoDB connection for a given database
and collection. If the connection does not exist, it creates a new one
using the `mongolite` package and stores it in a private environment
for reuse.
}
\examples{
\dontrun{
# Connect to the "users" collection in the "mydb" database
conn <- get_mongo_conn("mydb", "users", "mongodb://localhost:27017")
conn$find('{}')  # Query all documents
}

}
\seealso{
[mongolite::mongo()]
}
