% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{retry_read_rds}
\alias{retry_read_rds}
\title{Retry reading RDS file}
\usage{
retry_read_rds(file, max_tries = 3, delay = 0.3)
}
\arguments{
\item{file}{Path to the RDS file.}

\item{max_tries}{Number of retry attempts (default = 3).}

\item{delay}{Delay between retries in seconds (default = 0.3).}
}
\value{
The object read from the RDS file.
}
\description{
Attempts to read an RDS file multiple times before failing.
}
