% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{retry_save_rds}
\alias{retry_save_rds}
\title{Retry saveRDS with Delay and Limited Attempts}
\usage{
retry_save_rds(object, file, max_tries = 3, delay = 0.3)
}
\arguments{
\item{object}{An R object to save.}

\item{file}{A character string naming the file to save the R object to.}

\item{max_tries}{Maximum number of attempts before giving up. Default is 3.}

\item{delay}{Time in seconds to wait between attempts. Default is 0.3.}
}
\value{
`invisible(TRUE)` if successful. If all attempts fail, an error is raised.
}
\description{
Attempts to save an R object to a file using `saveRDS()`,
retrying on error up to a maximum number of tries.
This is useful in scenarios with high concurrency or
potential file system contention, such as APIs or parallel processes.
}
\examples{
\dontrun{
retry_save_rds(my_data, "config.rds")
}

}
