% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_to_mongo}
\alias{save_to_mongo}
\title{Save an R Object to MongoDB}
\usage{
save_to_mongo(
  user_id,
  object,
  file_name,
  db,
  uri,
  collection = Sys.getenv("ACCOUNT_BASE_DIR", "user_accounts")
)
}
\arguments{
\item{user_id}{A string specifying the unique user ID.}

\item{object}{The R object to save.}

\item{file_name}{The name of the file to save (e.g., `"account_tree.Rds"`).}

\item{db}{The MongoDB database name.}

\item{uri}{The MongoDB connection URI.}

\item{collection}{The MongoDB collection name. Defaults to `"user_accounts"`.}
}
\value{
Invisibly returns `NULL`. The object is stored in MongoDB.
}
\description{
This function saves an R object to a MongoDB collection for a specific user.
If a document with the same user ID and file name already exists, it updates
only the data field. Otherwise, it inserts a new document with metadata.
}
\examples{
\dontrun{
save_to_mongo(
  user_id = "user123",
  object = list(a = 1, b = 2),
  file_name = "account_tree.Rds",
  db = "mydb",
  uri = "mongodb://localhost:27017"
  collection = "user_accounts"
)
}

}
\seealso{
[get_mongo_conn()], [file_exists_mongo()], [remove_from_mongo()]
}
